+++ src/Fairoots/PluginInfo.cs
@@ -5,6 +5,6 @@ namespace Fairoots
     {
         public const string Guid = "OnlyCook.Fairoots";
         public const string Name = "Fairoots";
-        public const string Version = "1.0.1";
+        public const string Version = "1.0.2";
     }
 }